requirements.txt.txt
python-telegram-bot==20.0
matplotlib
pandas
